refact: implements MVC
