+++ nova-arquitetura-etapa-2/src/main/java/models/GenerateMusicResponse.java
@@ -1,6 +1,5 @@
-package controller.response;
+package models;
 
-import mapper.KeyMapper;
 import org.jfugue.player.Player;
 
 public class GenerateMusicResponse {
@@ -9,24 +8,23 @@ public class GenerateMusicResponse {
     private String textMusic = "";
 
     public GenerateMusicResponse(String text) {
-
         this.text = text;
     }
 
-    public void textProcessing () {
+    public void textProcessing() {
         // mapeamento para conseguir os caracteres reconhecidos pelo JFugue
         KeyMapper mapping = new KeyMapper();
         Player player = new Player();
-//        player.play("T290 I30 F F F F A# A# A# A# G# G# G# G# C# C# C# C#");
-//        System.out.println("aqui ?");
+        // player.play("T290 I30 F F F F A# A# A# A# G# G# G# G# C# C# C# C#");
+        // System.out.println("aqui ?");
 
-        for(char ch: this.text.toCharArray()) {
+        for (char ch : this.text.toCharArray()) {
             String jfugueKey = mapping.getMusicalAction((Character.toString(ch)));
-//            System.out.println(jfugueKey);
+            // System.out.println(jfugueKey);
             textMusic += jfugueKey + " ";
         }
 
-//        System.out.println(textMusic);
+        // System.out.println(textMusic);
         player.play(textMusic);
     }
 }
