+++ nova-arquitetura-etapa-2/src/main/java/models/enums/KeyLookup.java
@@ -1,4 +1,4 @@
-package enums;
+package models.enums;
 
 public enum KeyLookup {
