--- nova-arquitetura-etapa-2/src/main/java/service/InstructionsInterfaceService.java
@@ -1,4 +1,0 @@
-package service;
-
-public class InstructionsInterfaceService {
-}
