--- trabalho-pratico/src/main/java/Classes/ProgramRunner.java
@@ -1,11 +1,0 @@
-package main.java.Classes;
-
-/**
- * Executem essa classe para rodarem o programa inteiro.
- */
-
-public class ProgramRunner {
-    public static void main(String[] args) {
-        new UserInterface();
-    }
-}
